+++ src/main/java/io/enderdev/endermodpacktweaks/utils/EmtRender.java
@@ -1,0 +1,74 @@
+package io.enderdev.endermodpacktweaks.utils;
+
+import net.minecraft.client.renderer.BufferBuilder;
+import net.minecraft.client.renderer.GlStateManager;
+import net.minecraft.client.renderer.Tessellator;
+import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
+import org.lwjgl.opengl.GL11;
+
+import java.awt.*;
+
+public class EmtRender {
+    // <https://github.com/tttsaurus/Ingame-Info-Reborn/blob/4e79541b1f77ecf44f7d3718fd32555ca498d1cb/src/main/java/com/tttsaurus/ingameinfo/common/api/render/RenderUtils.java#L169>
+    public static void renderRoundedRect(float x, float y, float width, float height, float radius, Color color) {
+        int segments = Math.max(3, (int) (radius / 2f));
+        float a = (float) color.getAlpha() / 255.0F;
+        float r = (float) color.getRed() / 255.0F;
+        float g = (float) color.getGreen() / 255.0F;
+        float b = (float) color.getBlue() / 255.0F;
+
+        GlStateManager.glPolygonMode(GL11.GL_FRONT_AND_BACK, GL11.GL_FILL);
+        GL11.glEnable(GL11.GL_POLYGON_SMOOTH);
+
+        GlStateManager.disableCull();
+        GlStateManager.disableTexture2D();
+        GlStateManager.enableBlend();
+        GlStateManager.disableAlpha();
+        GlStateManager.tryBlendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
+        GlStateManager.color(r, g, b, a);
+
+        GlStateManager.glPolygonMode(GL11.GL_FRONT_AND_BACK, GL11.GL_FILL);
+        GL11.glEnable(GL11.GL_POLYGON_SMOOTH);
+
+        GlStateManager.pushMatrix();
+        GlStateManager.translate(0, 0, 0);
+
+        Tessellator tessellator = Tessellator.getInstance();
+        BufferBuilder bufferbuilder = tessellator.getBuffer();
+        bufferbuilder.begin(GL11.GL_TRIANGLE_FAN, DefaultVertexFormats.POSITION);
+
+        addArcVertices(bufferbuilder, x + width - radius, y + radius, radius, 0, 90, segments);
+        bufferbuilder.pos(x + width, y + radius, 0).endVertex();
+        bufferbuilder.pos(x + width, y + height - radius, 0).endVertex();
+        addArcVertices(bufferbuilder, x + width - radius, y + height - radius, radius, 90, 180, segments);
+        bufferbuilder.pos(x + width - radius, y + height, 0).endVertex();
+        bufferbuilder.pos(x + radius, y + height, 0).endVertex();
+        addArcVertices(bufferbuilder, x + radius, y + height - radius, radius, 180, 270, segments);
+        bufferbuilder.pos(x, y + height - radius, 0).endVertex();
+        bufferbuilder.pos(x, y + radius, 0).endVertex();
+        addArcVertices(bufferbuilder, x + radius, y + radius, radius, 270, 360, segments);
+        bufferbuilder.pos(x + radius, y, 0).endVertex();
+        bufferbuilder.pos(x + width - radius, y, 0).endVertex();
+
+        tessellator.draw();
+
+        GlStateManager.popMatrix();
+        GL11.glDisable(GL11.GL_POLYGON_SMOOTH);
+    }
+
+    private static void addArcVertices(BufferBuilder bufferbuilder, float cx, float cy, float radius, float startAngle, float endAngle, int segments) {
+        startAngle -= 90;
+        endAngle -= 90;
+        float x = (float) (cx + Math.cos(Math.toRadians(startAngle)) * radius);
+        float y = (float) (cy + Math.sin(Math.toRadians(startAngle)) * radius);
+        for (int i = 1; i <= segments; i++) {
+            float angle = (float) Math.toRadians(startAngle + (endAngle - startAngle) * i / segments);
+            float dx = (float) (cx + Math.cos(angle) * radius);
+            float dy = (float) (cy + Math.sin(angle) * radius);
+            bufferbuilder.pos(x, y, 0).endVertex();
+            bufferbuilder.pos(dx, dy, 0).endVertex();
+            x = dx;
+            y = dy;
+        }
+    }
+}